added navigation, added fragmentmanager, cleaned out main activity, added exchangeFragment
