+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph"
+    app:startDestination="@id/exchange_fragment">
+
+    <fragment
+        android:id="@+id/exchange_fragment"
+        android:name="com.example.changex.exchange_fragment"
+        android:label="exchange_layout"
+        tools:layout="@layout/exchange_layout" />
+</navigation>
