+++ app/src/main/res/values/strings.xml
@@ -40,4 +40,7 @@
         <item>Turkey</item>
     </string-array>
 
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+
 </resources>
